refactor to use a shell for top-level layout with child routes
