+++ src/Submissions.tsx
@@ -1,0 +1,9 @@
+function Submissions() {
+  return (
+    <section>
+      <h1 className="font-sans-lg">File Submissions</h1>
+    </section>
+  );
+}
+
+export default Submissions;
