+++ tests/Shell.test.tsx
@@ -1,5 +1,5 @@
 import { screen, render } from "@testing-library/react";
-import Dashboard from "../src/Dashboard";
+import Shell from "../src/Shell";
 import { vi } from "vitest";
 import {
   createMockedAuthContext,
@@ -9,7 +9,7 @@ import {
 
 vi.mock("react-oidc-context");
 
-describe("Dashboard", () => {
+describe("Shell", () => {
   afterEach(() => {
     vi.clearAllMocks();
     vi.restoreAllMocks();
@@ -18,7 +18,7 @@ describe("Dashboard", () => {
   it("should show sidebar", () => {
     render(
       withMockedAuthProvider(
-        withMemoryRouter(<Dashboard />, "/dashboard", { protected: true }),
+        withMemoryRouter(<Shell />, "/dashboard", { protected: true }),
         createMockedAuthContext({ isAuthenticated: true, isLoading: false })
       )
     );
@@ -29,7 +29,7 @@ describe("Dashboard", () => {
   it("should not show sidebar when not authenticated", () => {
     render(
       withMockedAuthProvider(
-        withMemoryRouter(<Dashboard />, "/dashboard", { protected: true }),
+        withMemoryRouter(<Shell />, "/", { protected: true }),
         createMockedAuthContext({ isAuthenticated: false, isLoading: false })
       )
     );
